Add files via upload
This is the code book for the file justmean.rda  That file is a data file which represents the mean of the mean and standard deviation variables of both the test and train datasets.  The data file itself is in the zip archive run_analysis_results.zip which is the result of running the R script run_analysis.R on the UCI HAR dataset.  The original URI for this dataset was "https://d396qusza40orc.cloudfront.net/getdata%2Fprojectfiles%2FUCI%20HAR%20Dataset.zip"